Add files via upload
## IMPORTANT INFO FOR .bsp.bz2 FILES ##
It's important to know that .bsp.bz2 files are too big for actual download, so they have been compressed into a .bz2 format. In order to access the simplified .bsp file, you're going to want to extract the files into a destination/new folder of your choice and then drag/drop the .bsp file into the recommended file path. See the readme.txt for more information on how to do that.